+++ src/NumSharp.Core/Statistics/np.nanvar.cs
@@ -143,7 +143,7 @@ namespace NumSharp
             var r = NDArray.Scalar(result);
             if (keepdims)
             {
-                var keepdimsShape = new int[arr.ndim];
+                var keepdimsShape = new long[arr.ndim];
                 for (int i = 0; i < arr.ndim; i++)
                     keepdimsShape[i] = 1;
                 r.Storage.Reshape(new Shape(keepdimsShape));
@@ -196,7 +196,7 @@ namespace NumSharp
                 for (int outIdx = 0; outIdx < outputSize; outIdx++)
                 {
                     // Convert flat index to coordinates in output shape
-                    var outCoords = new int[outputShape.Length];
+                    var outCoords = new long[outputShape.Length];
                     int temp = outIdx;
                     for (int i = outputShape.Length - 1; i >= 0; i--)
                     {
@@ -209,7 +209,7 @@ namespace NumSharp
                     int count = 0;
                     for (int k = 0; k < axisLen; k++)
                     {
-                        var inCoords = new int[inputShape.Length];
+                        var inCoords = new long[inputShape.Length];
                         int outCoordIdx = 0;
                         for (int i = 0; i < inputShape.Length; i++)
                         {
@@ -239,7 +239,7 @@ namespace NumSharp
                         double sumSq = 0.0;
                         for (int k = 0; k < axisLen; k++)
                         {
-                            var inCoords = new int[inputShape.Length];
+                            var inCoords = new long[inputShape.Length];
                             int outCoordIdx = 0;
                             for (int i = 0; i < inputShape.Length; i++)
                             {
@@ -269,7 +269,7 @@ namespace NumSharp
 
                 for (int outIdx = 0; outIdx < outputSize; outIdx++)
                 {
-                    var outCoords = new int[outputShape.Length];
+                    var outCoords = new long[outputShape.Length];
                     int temp = outIdx;
                     for (int i = outputShape.Length - 1; i >= 0; i--)
                     {
@@ -282,7 +282,7 @@ namespace NumSharp
                     int count = 0;
                     for (int k = 0; k < axisLen; k++)
                     {
-                        var inCoords = new int[inputShape.Length];
+                        var inCoords = new long[inputShape.Length];
                         int outCoordIdx = 0;
                         for (int i = 0; i < inputShape.Length; i++)
                         {
@@ -312,7 +312,7 @@ namespace NumSharp
                         double sumSq = 0.0;
                         for (int k = 0; k < axisLen; k++)
                         {
-                            var inCoords = new int[inputShape.Length];
+                            var inCoords = new long[inputShape.Length];
                             int outCoordIdx = 0;
                             for (int i = 0; i < inputShape.Length; i++)
                             {
@@ -340,7 +340,7 @@ namespace NumSharp
             // Handle keepdims
             if (keepdims)
             {
-                var keepdimsShapeDims = new int[arr.ndim];
+                var keepdimsShapeDims = new long[arr.ndim];
                 int srcIdx = 0;
                 for (int i = 0; i < arr.ndim; i++)
                 {
